+++ lib/src/bindings/bindings.dart
@@ -4,6 +4,8 @@ import "dart:io" show Platform;
 import "signatures.dart";
 import "structs.dart";
 
+// ignore_for_file: non_constant_identifier_names
+
 // bundles all C functions to be exposed to Dart
 class _ObjectBoxBindings {
   DynamicLibrary objectbox;
@@ -11,14 +13,14 @@ class _ObjectBoxBindings {
   // common functions
   void Function(Pointer<Int32> major, Pointer<Int32> minor, Pointer<Int32> patch) obx_version;
   Pointer<Uint8> Function() obx_version_string;
-  void Function(Pointer<Uint64> structPtr) obx_bytes_array_free,
-      obx_string_array_free,
+  void Function(Pointer<Uint64> structPtr) obx_string_array_free,
       obx_int64_array_free,
       obx_int32_array_free,
       obx_int16_array_free,
       obx_int8_array_free,
       obx_double_array_free,
       obx_float_array_free;
+  obx_free_t<OBX_bytes_array> obx_bytes_array_free;
   obx_free_t<OBX_id_array> obx_id_array_free;
 
   // error info
@@ -30,6 +32,8 @@ class _ObjectBoxBindings {
   // schema model creation
   Pointer<Void> Function() obx_model;
   int Function(Pointer<Void> model) obx_model_free;
+  int Function(Pointer<Void> model) obx_model_error_code;
+  Pointer<Uint8> Function(Pointer<Void> model) obx_model_error_message;
   int Function(Pointer<Void> model, Pointer<Uint8> name, int entity_id, int entity_uid) obx_model_entity;
   int Function(Pointer<Void> model, Pointer<Uint8> name, int type, int property_id, int property_uid)
       obx_model_property;
@@ -54,18 +58,19 @@ class _ObjectBoxBindings {
   int Function(Pointer<Void> txn) obx_txn_close;
   int Function(Pointer<Void> txn) obx_txn_abort;
   int Function(Pointer<Void> txn) obx_txn_success;
+  int Function(Pointer<Void> txn, int was_successful) obx_txn_mark_success;
 
   // box management
   Pointer<Void> Function(Pointer<Void> store, int entity_id) obx_box;
   int Function(Pointer<Void> box, int id, Pointer<Uint8> out_contains) obx_box_contains;
   int Function(Pointer<Void> box, Pointer<OBX_id_array> ids, Pointer<Uint8> out_contains) obx_box_contains_many;
-  int Function(Pointer<Void> box, int id, Pointer<Pointer<Void>> data, Pointer<Int32> size) obx_box_get;
-  Pointer<Uint64> Function(Pointer<Void> box, Pointer<OBX_id_array> ids) obx_box_get_many;
-  Pointer<Uint64> Function(Pointer<Void> box) obx_box_get_all;
+  int Function(Pointer<Void> box, int id, Pointer<Pointer<Uint8>> data, Pointer<IntPtr> size) obx_box_get;
+  Pointer<OBX_bytes_array> Function(Pointer<Void> box, Pointer<OBX_id_array> ids) obx_box_get_many;
+  Pointer<OBX_bytes_array> Function(Pointer<Void> box) obx_box_get_all;
   int Function(Pointer<Void> box, int id_or_zero) obx_box_id_for_put;
   int Function(Pointer<Void> box, int count, Pointer<Uint64> out_first_id) obx_box_ids_for_put;
-  int Function(Pointer<Void> box, int id, Pointer<Void> data, int size, int mode) obx_box_put;
-  int Function(Pointer<Void> box, Pointer<Uint64> objects, Pointer<Uint64> ids, int mode) obx_box_put_many;
+  int Function(Pointer<Void> box, int id, Pointer<Uint8> data, int size, int mode) obx_box_put;
+  int Function(Pointer<Void> box, Pointer<OBX_bytes_array> objects, Pointer<Uint64> ids, int mode) obx_box_put_many;
   int Function(Pointer<Void> box, int id) obx_box_remove;
   int Function(Pointer<Void> box, Pointer<Uint64> removed) obx_box_remove_all;
   int Function(Pointer<Void> box, Pointer<OBX_id_array> ids, Pointer<Uint64> removed) obx_box_remove_many;
@@ -122,6 +127,7 @@ class _ObjectBoxBindings {
 
   obx_query_visit_dart_t obx_query_visit;
 
+<<<<<<< HEAD
   // query property
   obx_query_prop_t<int> obx_query_prop;
   obx_query_prop_close_t<int> obx_query_prop_close;
@@ -145,6 +151,11 @@ class _ObjectBoxBindings {
   obx_query_prop_find_t<OBX_int8_array, Int8> obx_query_prop_int8_find;
   obx_query_prop_find_t<OBX_double_array, Double> obx_query_prop_double_find;
   obx_query_prop_find_t<OBX_float_array, Float> obx_query_prop_float_find;
+=======
+  // Utilities
+  obx_bytes_array_t<int> obx_bytes_array;
+  obx_bytes_array_set_t<int, int> obx_bytes_array_set;
+>>>>>>> dev
 
   // TODO return .asFunction() -> requires properly determined static return type
   Pointer<NativeFunction<T>> _fn<T extends Function>(String name) {
@@ -157,7 +168,9 @@ class _ObjectBoxBindings {
       libName += ".dll";
     } else if (Platform.isMacOS) {
       libName = "lib" + libName + ".dylib";
-    } else if (Platform.isLinux || Platform.isAndroid) {
+    } else if (Platform.isAndroid) {
+      libName = "lib" + libName + "-jni.so";
+    } else if (Platform.isLinux) {
       libName = "lib" + libName + ".so";
     } else {
       throw Exception("unsupported platform detected");
@@ -167,7 +180,7 @@ class _ObjectBoxBindings {
     // common functions
     obx_version = _fn<obx_version_native_t>("obx_version").asFunction();
     obx_version_string = _fn<obx_version_string_native_t>("obx_version_string").asFunction();
-    obx_bytes_array_free = _fn<obx_free_struct_native_t>("obx_bytes_array_free").asFunction();
+    obx_bytes_array_free = _fn<obx_free_t<OBX_bytes_array>>("obx_bytes_array_free").asFunction();
     obx_id_array_free = _fn<obx_free_t<OBX_id_array>>("obx_id_array_free").asFunction();
     obx_string_array_free = _fn<obx_free_struct_native_t>("obx_string_array_free").asFunction();
     obx_int64_array_free = _fn<obx_free_struct_native_t>("obx_int64_array_free").asFunction();
@@ -186,6 +199,8 @@ class _ObjectBoxBindings {
     // schema model creation
     obx_model = _fn<obx_model_native_t>("obx_model").asFunction();
     obx_model_free = _fn<obx_model_free_native_t>("obx_model_free").asFunction();
+    obx_model_error_code = _fn<obx_model_error_code_native_t>("obx_model_error_code").asFunction();
+    obx_model_error_message = _fn<obx_model_error_message_native_t>("obx_model_error_message").asFunction();
     obx_model_entity = _fn<obx_model_entity_native_t>("obx_model_entity").asFunction();
     obx_model_property = _fn<obx_model_property_native_t>("obx_model_property").asFunction();
     obx_model_property_flags = _fn<obx_model_property_flags_native_t>("obx_model_property_flags").asFunction();
@@ -209,6 +224,7 @@ class _ObjectBoxBindings {
     obx_txn_close = _fn<obx_txn_close_native_t>("obx_txn_close").asFunction();
     obx_txn_abort = _fn<obx_txn_abort_native_t>("obx_txn_abort").asFunction();
     obx_txn_success = _fn<obx_txn_success_native_t>("obx_txn_success").asFunction();
+    obx_txn_mark_success = _fn<obx_txn_mark_success_native_t>("obx_txn_mark_success").asFunction();
 
     // box management
     obx_box = _fn<obx_box_native_t>("obx_box").asFunction();
@@ -317,6 +333,10 @@ class _ObjectBoxBindings {
     obx_query_prop_int8_find = _fn<obx_query_prop_find_t<OBX_int8_array, Int8>>("obx_query_prop_int8_find").asFunction();
     obx_query_prop_double_find = _fn<obx_query_prop_find_t<OBX_double_array, Double>>("obx_query_prop_double_find").asFunction();
     obx_query_prop_float_find = _fn<obx_query_prop_find_t<OBX_float_array, Float>>("obx_query_prop_float_find").asFunction();
+
+    // Utilities
+    obx_bytes_array = _fn<obx_bytes_array_t<IntPtr>>("obx_bytes_array").asFunction();
+    obx_bytes_array_set = _fn<obx_bytes_array_set_t<Int32, IntPtr>>("obx_bytes_array_set").asFunction();
   }
 }